requirements.txt
selenium
webdriver-manager
python-dotenv
beautifulsoup4
requests
python-telegram-bot